Move ProvisioningParameters to common second part
